Bug 989069 - Smooth out the API of nsJSUtils::EvaluateString. v1
Only one callers wants the funny convert-exception-to-return-value behavior, and
it can just handle that on its own.

Note that, historically, in the reportUncaught=true cases (the default), the only
time we'd return something other than NS_OK is in the case that JS_WrapValue
failed. In particular, we'd return true even if script threw an exception. So we
want to be careful not to break callers (especially XBL field installation) that
might be accustomed to pushing throw exceptions. Doing this the simple way means
that we don't propagate the JS_WrapValue OOMs in all the cases, but this definitely
an improvement.