Bug 1062631 - Fix up content QI security check and tests. v1 r=bz
This stuff is broken in a bunch of ways. The test isn't doing the right thing
because it's using SpecialPowers. When we pass the content object to
xpconnectArgument(), it gets wrapped in a SpecialPowers wrapper, so the check
succeeds, only to fail down the line when the QI method is actually invoked
with an IID (which we can't pass to content scopes anyway).

This whole security check is probably useless given the above, but let's be
safe, make it a bit more robust, and fix up the test while we're at it.